Create license
Hello. Thanks for the Keras implementation. I noticed the number of forks and thought you might want to make explicit the kinds of usage you would like to allow with a license doc. This MIT license is common and straightforward but of course you should use whatever you think is appropriate.